feat(ai): migrate from OpenRouter to GitHub Models API
- Replace OpenRouter AI service with GitHub Models API
- Implement Llama-3.2-11B-Vision-Instruct model integration
- Add comprehensive AI assistant functionality:
  - Personalized farming advice
  - Anonymous farming advice
  - Crop disease diagnosis
  - Weekly farming tips generation
- Update environment variables from OPENROUTER_API_KEY to GITHUB_TOKEN
- Remove OpenAI dependency, add node-fetch for API calls
- Update all AI routes to use new GitHub Models service
- Fix TypeScript errors and add proper type definitions
- Update test mocks to use new service
- Add GitHub Models setup documentation
- Update README.md and API_DOCUMENTATION.md with new features
- All tests passing (38 test cases)
- Free AI service with no cost limitations

Breaking Changes:
- OPENROUTER_API_KEY environment variable replaced with GITHUB_TOKEN
- OpenRouter AI service removed in favor of GitHub Models
