+++ package.json
@@ -29,7 +29,7 @@
     "helmet": "^8.1.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.16.4",
-    "openai": "^5.10.1",
+    "node-fetch": "^3.3.2",
     "winston": "^3.17.0"
   },
   "devDependencies": {
@@ -40,6 +40,7 @@
     "@types/jest": "^30.0.0",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^24.0.14",
+    "@types/node-fetch": "^2.6.12",
     "@types/nodemon": "^1.19.6",
     "@types/supertest": "^6.0.3",
     "@typescript-eslint/eslint-plugin": "^8.37.0",
